+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - The account balance no longer runs around like a slot machine while loading transactions.
+- Months now sort correctly when switching between account views.
 
 ## [0.16.8] - 2023-03-01
 ### Changed